requirements.txt
kaldiio
numpy
matplotlib
configargparse
tensorboard
editdistance
PyYAML
jiwer
sentencepiece
soundfile
librosa
torch_complex
